+++ changes.md
@@ -4,3 +4,4 @@
   - Adds tests for these new classes.
 - Adds overhead class which accounts for pre- and post- overhead timings for both VISDA and NIRDA
 - Adds baseline short-term calendar runner script to docs/
+- Adds ability to merge back-to-back observations of the same target